+++ anton/core/memory/cerebellum.py
@@ -70,7 +70,7 @@ from typing import TYPE_CHECKING
 from pydantic import BaseModel, Field
 
 from anton.core.backends.base import Cell
-from anton.core.memory.hippocampus import Engram
+from anton.core.memory.base import Engram
 
 if TYPE_CHECKING:
     from anton.core.llm.client import LLMClient
